Add files via upload
Generate the HTML for a link that opens in a separate browser tab. The attribute values of the link should be in double quotes. Generate the link 2 different ways. First, use double quoted strings with embedded double quotes, then use single quoted strings with embedded double quotes. Now use arrays each with a handful of elements and loop through the arrays to generate links to multiple pages. Do the same to generate multiple images.