Add files via upload
Write a function that takes 2 parms, a string and a character. Return a string where the character passed to the function is placed in between every character of the string. For example, if you pass "testing" and "*" to the function, the function will return "t*e*s*t*i*n*g*". Rewrite the previous function so that if you pass "testing" and "*", the function will return "t*e*s*t*i*n*g".